+++ mapentity/forms.py
@@ -85,11 +85,12 @@ class TranslatedModelForm(forms.ModelForm):
 class SubmitButton(HTML):
 
     def __init__(self, div_id, label):
+        # onclik is here to bypass html5 validation
         content = ("""
-            <a id="{0}" class="btn btn-success"
-               onclick="javascript:$(this).parents('form').submit();">
+            <button type="submit" id="{0}" class="btn btn-success"
+                   onclick="this.form.submit();">
                 <i class="bi bi-check-circle-fill"></i> {1}
-            </a>""".format(div_id, label))
+            </button>""".format(div_id, label))
         super().__init__(content)